# use non empty cells with the reduce_cell_operator operator