This PR is an optimization of the reduce_cell_particles operator for simulation with a lot of empty cells (90% or more). Speedup=7.2 for my example, two reductions -> vertex_displ and simulation_state.

Old results : 

```
  loop ............................................  1.304e+04            0.000   0.000         1  99.68% / 100.00%
    scheme ........................................  6.127e+03            0.000   0.000     10000  46.84% / 46.98%
      combined_compute_prolog .....................  1.688e+02            0.000   0.000     10000   1.29% /  1.29%
      mpi_barrier
      check_and_update_particles ..................  5.083e+03            0.000   0.000     10000  38.86% / 38.98%
        vertex_displ_over .........................  4.854e+03            0.000   0.000     10000  37.10% / 37.22%

...

    end_iteration .................................  6.904e+03            0.000   0.000     10000  52.78% / 52.95%
      nth_timestep
      nth_timestep
      nth_timestep
      nth_timestep
      nth_timestep
      trigger_thermo_state ........................  3.145e+00            0.000   0.000     10000   0.02% /  0.02%
        boolean_or
        boolean_or
      nth_timestep
      thermo_state_if_triggered ...................  6.877e+03            0.000   0.000      1001  52.57% / 52.74%
        simulation_state ..........................  6.876e+03            0.000   0.000      1001  52.57% / 52.73%
```

Now 

```

  loop ............................................  1.820e+03            0.000   0.000         1  98.15% / 100.00%
    scheme ........................................  1.503e+03            0.000   0.000     10000  81.06% / 82.59%
      combined_compute_prolog .....................  1.490e+02            0.000   0.000     10000   8.04% /  8.19%
      mpi_barrier .................................  5.390e-01            0.000   0.000     10000   0.03% /  0.03%
      check_and_update_particles ..................  5.895e+02            0.000   0.000     10000  31.80% / 32.40%
        vertex_displ_over .........................  3.992e+02            0.000   0.000     10000  21.53% / 21.94%
...

    end_iteration .................................  2.342e+02            0.000   0.000     10000  11.33% / 11.54%
      nth_timestep ................................  8.717e-01            0.000   0.000     10000   0.04% /  0.04%
      nth_timestep ................................  8.892e-01            0.000   0.000     10000   0.04% /  0.04%
      nth_timestep ................................  8.875e-01            0.000   0.000     10000   0.04% /  0.04%
      nth_timestep ................................  6.498e-01            0.000   0.000     10000   0.03% /  0.03%
      nth_timestep ................................  6.577e-01            0.000   0.000     10000   0.03% /  0.03%
      trigger_thermo_state ........................  2.498e+00            0.000   0.000     10000   0.12% /  0.12%
        boolean_or ................................  4.440e-01            0.000   0.000     10000   0.02% /  0.02%
        boolean_or ................................  4.345e-01            0.000   0.000     10000   0.02% /  0.02%
      nth_timestep ................................  7.233e-01            0.000   0.000     10000   0.03% /  0.04%
      thermo_state_if_triggered ...................  2.109e+02            0.000   0.000      1001  10.20% / 10.40%
        simulation_state ..........................  2.107e+02            0.000   0.000      1001  10.19% / 10.38%
```